Javadoc + Test

+++ API_Order/src/main/java/fr/univamu/iut/api_order/Order.java
@@ -1,7 +1,5 @@
 package fr.univamu.iut.api_order;
 
-import java.util.Date;
-
 /**
  * Classe représantant une Commande
  */
@@ -15,7 +13,7 @@ public class Order {
     /**
      * Prix de la commande
      */
-    protected float orderPrice;
+    protected int orderPrice;
 
     /**
      * Localisation point de relais
@@ -28,22 +26,30 @@ public class Order {
     protected String orderDate;
 
     /**
-     *
+     * Validation
      */
     protected boolean isValidate;
 
     /**
-     *
+     * liste des produits
      */
     protected String recapOrder;
 
     /**
-     *
+     * Constructeur par défaut
      */
     public Order(){}
 
-
-    public Order(int idOrder, float orderPrice, String relayPlace, String orderDate, boolean isValidate, String recapOrder) {
+    /**
+     * Constructeur
+     * @param idOrder
+     * @param orderPrice
+     * @param relayPlace
+     * @param orderDate
+     * @param isValidate
+     * @param recapOrder
+     */
+    public Order(int idOrder, int orderPrice, String relayPlace, String orderDate, boolean isValidate, String recapOrder) {
         this.idOrder = idOrder;
         this.orderPrice = orderPrice;
         this.relayPlace = relayPlace;
@@ -52,56 +58,72 @@ public class Order {
         this.recapOrder = recapOrder;
     }
 
+    /**
+     * Modifie la validation du commande
+     * @param validate
+     */
     public void setValidate(boolean validate) {
         isValidate = validate;
     }
 
+    /**
+     * Vérifie si la commande est validée ou non
+     * @return boolean
+     */
     public boolean isValidate() {
         return isValidate;
     }
 
     /**
-     *
-     * @return
+     * Récupère l'ID de la commande
+     * @return int
      */
     public int getIdOrder() {
         return idOrder;
     }
 
+    /**
+     * Modifie la liste des produits de la commande
+     * @param recapOrder
+     */
     public void setRecapOrder(String recapOrder) {
         this.recapOrder = recapOrder;
     }
 
+    /**
+     * Récupère la liste des produits de la commande
+     * @return String
+     */
     public String getRecapOrder() {
         return recapOrder;
     }
 
     /**
-     *
-     * @return
+     * Récupère le prix total de la commande
+     * @return float
      */
-    public float getOrderPrice() {
+    public int getOrderPrice() {
         return orderPrice;
     }
 
     /**
-     *
-     * @return
+     * Récupère le relai de livraison de la commande
+     * @return String
      */
     public String getRelayPlace() {
         return relayPlace;
     }
 
     /**
-     *
-     * @return
+     * Récupère la date de la commande
+     * @return String
      */
     public String getOrderDate() {
         return orderDate;
     }
 
     /**
-     *
+     * Modifie l'Id de la commande
      * @param idOrder
      */
     public void setIdOrder(int idOrder) {
@@ -109,15 +131,15 @@ public class Order {
     }
 
     /**
-     *
+     * Modifie le prix total de la commande
      * @param orderPrice
      */
-    public void setOrderPrice(float orderPrice) {
+    public void setOrderPrice(int orderPrice) {
         this.orderPrice = orderPrice;
     }
 
     /**
-     *
+     * Modifie le relai de livraison de la commande
      * @param relayPlace
      */
     public void setRelayPlace(String relayPlace) {
@@ -125,13 +147,17 @@ public class Order {
     }
 
     /**
-     *
+     * Modifie la date de la commande
      * @param orderDate
      */
     public void setOrderDate(String orderDate) {
         this.orderDate = orderDate;
     }
 
+    /**
+     * Méthode toString
+     * @return String les informations d'une commande
+     */
     @Override
     public String toString() {
         return "Order{" +
